Remove devilspie.
Our Devilspie-configurations were intended for bigtouch mode.
Well, bigtouch is broken on Bookworm and it was not used pretty
much anywhere ever.  Thus, Devilspie configurations are set up
at boot for everyone and Devilspie is started at login for everyone,
even though it is used by no one.  Enough reasons to remove.

This may come back if needed but clearly some redesign is needed.
If bigtouch mode is to return, our bigtouch Gnome-extension should
probably startup Devilspie with the configurations it needs.