darkdm: Move system info lines around and print some extra info.
Important system info starts from the top,
additional info is more to the bottom.

Print the name of the current bootserver, this is useful
for remotely troubleshooting unregistered devices, as it
may be necessary to hop through the bootserver to access them.

Print "No license" if there's no Windows license in DMI table,
instead of leaving the line empty as previously.