Add "connect host to network"-functionality to reset mode.
Because resetting Windows requires network access (Windows-image
must be downloaded), and a device may not be on a network when
resetting, provide means to connect host to a network from the
reset mode.  Require network in this case and when Windows reset
is not needed do not bother people about it.