Do not include Grub code for choosing kernels when UKI kernels are used.
Those are pointless in that context (do not work), and Grub
configuration should be rewritten to use loopback only when
absolutely necessary (when there are no UKI kernels to boot from).
The reason for this is that loopback can not be used when
SecureBoot is used, Grub starts to measure the image to TPM
and fails to do so.