Add "puavo-install-uki-grub".
This is meant to handle the UKI Grub installation case.
This is "puavo-install-grub" but only contains things needed
in the UKI case.  As there share functionality, some kind of
merging or shared codebase is needed.

The "puavo-install-grub" might also become a wrapper that
does call the correct script underneath.  We'll see.