Include testipp in cups build and installation
Would have been very nice and straightforward change to just add the
0103-build-testipp.patch with the new puavo build/patch mechanism, but as the
binary also needs to be included in some of the debs created, the debian
cups-ipp-utils.install needed to be changed, too, and doing this with a patch
didn't seem to be possible, packaging scripts getting very confused by it.
So in addition to a couple of lines of Makefile patch, this commit also returns
cups build to use the old debian mechanism.