Disable PAC on Kerberos tickets.
I do not know if we have any use for it, I think not.
The PAC data breaks at least the Bullseye Samba version as we use
it, see https://bugzilla.samba.org/show_bug.cgi?id=15582.
That is, if Kerberos master provides a ticket with PAC set,
the Samba in Bullseye bootservers rejects it.  For now
we just disable PAC.