Link cups against gnutls after all, not libssl.
Changed my mind...  There is a certificate generation issue
with libssl (double base64-encoding), which means that CUPS
certificates generated with libssl-layer are not compatible
with anything else than CUPS, even including older versions
of CUPS or versions linked against gnutls.

I see that Fedora 40 links CUPS against gnutls and that is
apparently the intention of the Debian-packaging rules as well,
thus add "libssl-dev" to conflicting packages so that
gnutls-linkage will be effective.