Sync clock each time network is up, not only when puavo-vpn gets up.
This is to fix clock time in installation time.  The device
registering phase needs a clock time when our puavo-domain
certificates are valid, and these get renewed from time to time.
As the maintenance vpn is not available during that time,
we should sync clock also when network gets up.  The systemd-timesyncd
is shy to make bigger jumps in this situation, which is okay.