Do not start the exammode service in the background.
I think we should probably restart the exammode for each exam session.
This is because the temporary filesystem state remains from the
previous sessions if we try to run the exam sessions on the same
Xorg instance.

It is slower and not as beautiful, though.  Some fixup commits
are still needed.