Make it possible to disable background menu through gsettings/dconf.
Added the "/org/gnome/desktop/lockdown/disable-background-menu" key
for configuring this.  It is "false" by default, but is set to "true",
the background image menu is disabled.  Set this value to "true"
for the examination mode.  Using this is useful for the examination mode,
otherwise users can modify desktop settings and also run programs.