Cleanup unused things from puavo-exammode-session-applet.
This is actually not an applet, but might become one some day.
It does not present an interface to a user, no notifications,
no dbus communications, but it does handle the puavo-exam-browser
startup, and thus is needed.  This and the puavo-exammode gnome-shell
extension are responsible for most the exammode UI interface,
and we'll see how the responsibilities should be split up between
these two components.