Do not use mktemp when creating "defused Windows EFI"-tar.
"mktemp" is not needed for security because this directory is
writable only by root.  And the problem with randomly chosen temporary
files is that something should also clean those up in case of problems,
and hose will not be cleaned up automatically because this is not "/tmp".
One should not assume that given two commands in a sequence
the latter command gets to run.