puavo-manage-efi: require excplicit disk devpath as arg
Previously, puavo-manage-efi expected to find just one ESP, but it's
certainly possible that there are multiple ESPs on different disks
etc. Require the disk of the Windows installation explicitly as an
argument.

setup_grub_environment script assumes the Windows installation disk is
the same which contains /images. Which is quite reasonable assumption:
we only support single-disk installations for dual-boot hosts.