Do not run puavo-conf-update inside initramfs.
The idea behind this was that we could apply some hwquirks
in the initrd-phase.  After ten years of hindsight, this does
not look like it was worth it.  The puavo-conf-update has to
be run twice (because device.json in only available at the later
stage).

In the initrd-phase, do still look for the special puavo.* kernel
command line arguments and act accordingly.  Now the logic to
determine those has to exist outside initrd, but the functionality
(if it is still needed) is still there.

This seems like a sane way forward, but we'll see what comes out
of this.