Do not add loop= parameter to kernel command line and do not handle it at boot.
Our Trusty-based systems used this.  Removing this means
that Puavo devices with Trusty Grub configuration can not
boot Trixie systems (or newer), and Trixie-based systems can not
boot Trusty systems.

That said, it should still be possible to upgrade/downgrade
systems between these, even though I am not sure if that will
work or if we're even going to test that.  Trusty was EOL six
years ago and machines using that should have been upgraded
a long time ago.