debs/linux: Upgrade to 6.12.12-1~bpo12+1.debian
This commit upgrades linux package to 6.12.12-1~bpo12+1.debian, and applies
the Puavo derivative for Microsoft Surface patch on top.

The Surface patch is based 6.12 patch in the commit ID:

https://github.com/linux-surface/linux-surface/commit/0e17ba8007708620dd759cab3f107ea033514c43

The patch can be also applied as the last entry to the `series` file when
running self-compiled kernel in Debian (as long as the kernel compiled
from the same sources).

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@opinsys.com>