core/ltsp/rules: add puavo.dockerd.data-root-in-home configuration option
New configuration option puavo.dockerd.data-root-in-home is false by default.

When it's true, dockerd is configured to use /home/.dockerd-data-root
as it's data root directory (place for all images, container, volumes,
etc.).

The need/reason for this option is that docker images, containers,
volumes et al. require substantial amount of disk space, and /var
(which most often is just in the root partition), does not meet the
needs.

Configuration is done in preinit phase by writing to /etc/docker/daemon.json .