In puavo-setup-filesystems, use puavo-get-boot-disks to lookup the current boot devices.
This is to prevent installation to disks where the installer
has booted from (that can not work).  Also be stricter and
stop the installation process in case puavo-get-boot-disks
returns error.

This also now takes into account that the installation disk
could be spread as a RAID configuration over multiple disks.
No one should be doing something like that, but that should work
;-)