darkdm: Print WiFi 6 (in)compatibility note only for wlan*.
Previously it printed the note on whatever was the
last line printed via the list_networks function.
Didn't matter if it was wlan0, eth0 or from the new naming scheme.

To be fair, that was my fault, as that function was done
very quickly and in need, thus tidiness wasn't the priority. :D

It was recently brought to my attention that printing
"Not Wifi 6 compatible" can be a bit misleading
when a device actually doesn't have wifi at all.

So with this commit, that's now solved.