Just use do-debpool-changes from PATH in .aux/debs-build.
There are situations where the ../../ path traversal does not work.
For cases where /usr/bin/do-debpool-changes is not available or
is old, we could set PATH suitably and use the version in the
source tree.  This might be done in our Makefile or somewhere
(but not done in this commit).