Stop using "user_type" and start using "user_roles" everywhere.
The problem with "user_type" is that it is singular, that is,
through that user can have only one role at a time.  But Puavo
supports multiple roles per user.  The "user_roles" key provides
that information, but that has not been used thus far.

Now user roles should provide some permissions correctly in case
the user has several roles.