Remove the "ppolicy" module from the extldap configuration.
Enforcing password policy does not make much sense in a
read-only server, and we have already done this for bootservers
(that have a read-only ldap).  And, because the current ldapmaster
still contains a ppolicy schema that breaks the ppolicy module
(and thus slapd start entirely), it is probably safer to simply
turn this off.

The current plan is to keep the ppolicy module on the ldapmaster
still, even though we do not use it, but we possible might some
day.  (Puavo can already enforce some password validation).