Move logic around in puavo-disk-erase and puavo-disk-clone.
I did an oopsie and didn't realize the previous changes would break
puavo-disk-erase completely on sda-devices when running by itself.

Now that's fixed and it will fallback to blkdiscard whenever erasing
a disk while a certain realtek nic is present, no matter if
doing a restoredisk or just an erase.