Revert "Change the squashfs image compression algorithm to "zstd"."
This reverts commit 6a9f28611b6c5e1c510dcb780441a640bf9f527e.

Unfortunately this needs to be reverted.  While this works
on fresh Puavo OS installs using Trixie, that boot UKI kernels
on EFI partitions, our legacy installations look up kernel
from squashfs, and the Grub squashfs module currently
does not support zstd decompression.

The "xz" algorithm should be an option if we need to make
the image smaller, but it is slower, so go back to "lzo".

One option is to use a different image for legacy installations,
but that managing that nicely is a big problem in itself.