Add our modifications to OpenLDAP Debian package build rules.
I am leaving two customizations out to see if those are still
needed.  One is that we have downloaded ldap schemas from
www.openldap.org.  As this is a fifteen year old hack, I suspect
it will not be necessary.  The other thing is that I do not
think dropping the smbk5pwd module should be necessary, so
I am not accept that trick as is.  (Here I am referring to
old hacks we still have in Bookworm.)

The smbkrb5pwd module does not build yet, so that still needs
some fixes.