Build hooktftp with the "puavo"-mechanism.
This was surprisingly tricky, as there is no separate debian-packaging
rules tarball.  I also noticed, that the original source gets
unpacked twice, because debs-reset-debian removes the source
before it is created again.  This seems wrong, so tring without
the cleanup and let's see if other things break due to this.