OpenQA: pick the latest install image to boot from, by a different method.
Now that it sets up the images by symlinks, the symlink modification
time may not be the most reliable way to determine which image
is the latest.  Or perhaps it is.  But right now, as the
image filenames sort so that the last is the latest, pick that one.

I do not know how this should really be done.  Pick one at random?
Or always require an explicit choice?