OpenQA: Use two gigabytes of ram for test virtual machines.
Debian default was one gigabyte, which was not enough to boot
Puavo OS.  The unpacking of the initrd would fail due to
it not fitting into memory (!), and the system would just
end up in a black screen.

Odd!  I would have expected that Puavo OS could boot on
machines with only one gigabyte of memory!