Simplify/stupidify the puavo-ers-applet virtual machine shutdown sequence.
Perhaps it could observe it shutting down, but this should be
good enough (we should do automatic reboots only at night).
The old optimistic delays did not work, because shutting down
the virtual machine could take a minute, and then perhaps
the VirtualBox could inhibit a reboot (?).  This appears to do
the right things, only slowly.