Add reboot counter when installing system through darkdm.
This is (partially) gotten back from commit
d3418962a55431a99a9b2faee3d1cf3370e76938.

Reboot counter during installation is nice because in automatic testing
it should allow us some time to check the final installation messages
on the screen.

I am not reverting that old commit as is, though, because
when interactively doing a reboot or a poweroff, the counter
is annoying.