In btrfs quota setup, handle the case of image overlay in a special way.
The default imageoverlay of zero quota is good, because the imageoverlay
is not normally used.  However, when the image overlay is enabled,
we should also set the quota to a non-zero value so that things will
not break badly, and do that as soon as possible.  This commit hooks
the puavo.image.overlay puavo-conf variable to quota setup, which
is run as soon as the variable changes.   Also in case the quota
is set to zero and overlay is used, we choose five gigabyte quota
instead.