Add a puppet module, "benchmarks", for benchmarks.
Contains "puavo-benchmark" that current just runs the PassMark
Performance Test.  Include this by default in basic images.

I do not want to integrate this as part of the puavo-ltsp-client
or some package, because PassMark is not open source.  We choose
it because the Finnish Matriculation Examination Board uses this
test to determine if a computer has enough performance for
conducting their exams.