Include the /etc/systemd/network/99-default.link file (or link to /dev/null) in the Dracut image.
This is because we want the systemd-udevd to behave the same
regarding interface names both inside initrd and outside it.

Our policy has been to go with the kernel assigned interface
names, and special udev-rules can be used to make those
persistent.  This is partially due to historical reasons,
and if we want to change that, we should think about that
carefully (to not break upgrades to Trixie).