puavo-boot-trust-manager: Explicitly mount the root device
Previously, the system relied on the systemd-gpt-auto-generator for root
device discovery and LUKS setup. This approach is limited by the default
discovery timeout of 90 seconds. While this timeout can be adjusted, it
currently appears to require changing global device timeouts, which is
not ideal.

This commit shifts responsibility to the boot trust manager by defining
the root device in kernel parameters. This change aims to bypass the
short discovery window. A longer Dracut timeout still remains after this
change, though it is intended to be addressed in future commits.